src/scout/requirements.txt
pillow==9.3.0
requests==2.28.1
